refactor: move can_sign_add_signature to core
